Added running string

+++ Core/Src/NOS_LIB_Driver/Inc/WS2812B.h
@@ -7,7 +7,7 @@ void NOS_WS2812B_Matrix_Init(WS2812B_Matrix* matrix,uint8_t* buff,uint16_t ledsC
 
 void NOS_WS2812B_Matrix_SetPixel(WS2812B_Matrix* matrix,PixelColor* color,uint16_t pixelPos); //Set Pixel in Leds strip by pos number
 
-void NOS_WS2812B_Matrix_Update(WS2812B_Matrix* matrix,uint8_t offset); //Update frame. For stable work do it every 35-50 ms in while
+void NOS_WS2812B_Matrix_Update(WS2812B_Matrix* matrix,uint16_t offset); //Update frame. For stable work do it every 35-50 ms in while
 
 void NOS_WS2812B_Matrix_SetSymvol(WS2812B_Matrix* matrix,char sym,uint8_t pos);
